#16 Refactoring
