+++ pylasu/emf/__init__.py
@@ -1,0 +1,2 @@
+from .metamodel_builder import MetamodelBuilder
+from .model import *
